% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{drop_code_blocks}
\alias{drop_code_blocks}
\title{Drop the lines of fenced code blocks}
\usage{
drop_code_blocks(lines)
}
\arguments{
\item{lines}{A character vector.}
}
\value{
A character vector with fenced code blocks removed.
}
\description{
Internal function for \code{\link[=strip_markdown]{strip_markdown()}}.
}
\keyword{internal}
