% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{escape_html}
\alias{escape_html}
\title{Escape the characters that mean something in HTML}
\usage{
escape_html(text)
}
\arguments{
\item{text}{A character vector.}
}
\value{
A character vector.
}
\description{
Internal function for \code{\link[=html_sujimichi]{html_sujimichi()}}.
The ampersand goes first, or the ampersands of the other escapes
would be escaped in their turn.
}
\keyword{internal}
