% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text.R
\name{join_lines}
\alias{join_lines}
\alias{move_mark_after_close}
\title{Join the lines of a paragraph into one string}
\usage{
join_lines(text)

move_mark_after_close(text, mark)
}
\arguments{
\item{text}{A string.}

\item{mark}{A string used as the split mark.}
}
\value{
A string.
}
\description{
Internal function for \code{\link[=split_sentences]{split_sentences()}}.
}
\keyword{internal}
