% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_words.R
\name{morpheme_col}
\alias{morpheme_col}
\title{Take a column out of a morpheme table}
\usage{
morpheme_col(morphemes, name)
}
\arguments{
\item{morphemes}{A data.frame of morphemes.}

\item{name}{A string: "form", "pos", "pos_1" or "lemma".}
}
\value{
A vector, or \code{NULL} when the column is missing.
}
\description{
Internal function for \code{\link[=pick_content_words]{pick_content_words()}}.
Looks for the English column name first, then for the Japanese one
used by 'moranajp'.
}
\keyword{internal}
