% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_view.R
\name{overlaps}
\alias{overlaps}
\title{Tell whether one string holds the other}
\usage{
overlaps(a, b)
}
\arguments{
\item{a, b}{Strings.}
}
\value{
A logical.
}
\description{
Internal function for \code{\link[=mark_more]{mark_more()}}.
}
\keyword{internal}
