% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{strip_markdown_line}
\alias{strip_markdown_line}
\title{Strip the Markdown notation of one line}
\usage{
strip_markdown_line(line)
}
\arguments{
\item{line}{A string.}
}
\value{
A string.
}
\description{
Internal function for \code{\link[=strip_markdown]{strip_markdown()}}.
}
\keyword{internal}
